Refactor and Updated genome-wide over all contigs
Previously the genome-wide over all contigs plot filtered data based on min frequency, however the original R script did not do this, so I updated the plot to better reflect the R script's behavior. It is not clear if the original script's behavior is intentional but I'll leave it like this for now.